Produce only one Fernet key per image, instead of per container
Fernet is used to encrypt passwords in Connections, if the Fernet key is
different between executions of the local runner (where a new container
is created) the key used to encrypt the keys will be different than the
key used to decrypt. This commit creates one key during image build and
is re-used by all containers until rebuild.

fixes #27
fixes #24
closes #28