Add SELinux sharing labels
This fixes running the environment on a Linux system with enabled SELinux (such as Fedora 39). Note that I opted for
a more open :z instead of private sharing using :Z, mainly because I am not yet familiar with the exact setup and which
parts run together with which other parts.

See also https://prefetch.net/blog/2017/09/30/using-docker-volumes-on-selinux-enabled-servers/